+++ src/sql/base/common/locConstants.ts
